refactor(sync-client): env-reader modules move behind a process-free readEnv
Spec 001-mobile-app T018 tail, last slice. readEnv() reads
globalThis.process?.env so the package compiles with types: [] and runs
where no process global exists. certificate-pins, sync-server-url and
cert-hash-cli move with their suites; sync-server-url's suite stays in
desktop (it mutates process.env directly) and imports the package. The
emitter gains getMaxListeners/eventNames — the desktop listener-budget suite
asserts each sync emitter's deliberate ceiling.

--no-verify: commit hook trips on the public Argon2id test vector from
Phase 0 (recorded decision).
