+++ apps/desktop/src/main/canvas/assets/asset-service-context.test.ts
@@ -35,7 +35,7 @@ vi.mock('../../agent/storage/vault-id', () => ({ getOrCreateVaultUuid }))
 vi.mock('../../ipc/sync-attachment-handlers', () => ({ getCanvasAssetIO }))
 vi.mock('../../sync/token-manager', () => ({ getValidAccessToken }))
 vi.mock('../../sync/runtime', () => ({ getSyncEngine }))
-vi.mock('../../sync/sync-server-url', () => ({ resolveSyncServerUrl }))
+vi.mock('@memry/sync-client/sync-server-url', () => ({ resolveSyncServerUrl }))
 vi.mock('../../sync/crdt-writeback', () => ({ markWritebackIgnored }))
 vi.mock('../../telemetry/track', () => ({ trackMainEvent }))
 vi.mock('./attachment-dereference', () => ({ dereferenceChunks }))
