+++ apps/desktop/src/main/canvas/assets/asset-service-context.ts
@@ -15,7 +15,7 @@ import { getDatabase, isDatabaseInitialized } from '../../database/client'
 import { getCanvasAssetIO } from '../../ipc/sync-attachment-handlers'
 import { markWritebackIgnored } from '../../sync/crdt-writeback'
 import { getSyncEngine } from '../../sync/runtime'
-import { resolveSyncServerUrl } from '../../sync/sync-server-url'
+import { resolveSyncServerUrl } from '@memry/sync-client/sync-server-url'
 import { getValidAccessToken } from '../../sync/token-manager'
 import { trackMainEvent } from '../../telemetry/track'
 import { getStatus as getVaultStatus } from '../../vault/index'
