+++ apps/desktop/src/main/ipc/sync-attachment-handlers.ts
@@ -59,7 +59,7 @@ import {
 } from '../sync/note-attachment-metadata'
 import { registerCommand } from './lib/register-command'
 import { getNetworkMonitor } from '../sync/runtime'
-import { resolveSyncServerUrl } from '../sync/sync-server-url'
+import { resolveSyncServerUrl } from '@memry/sync-client/sync-server-url'
 import { getValidAccessToken } from '../sync/token-manager'
 import { getOrCreateVaultUuid } from '../agent/storage/vault-id'
 import { getMainI18n } from '../lib/main-i18n'
