+++ apps/desktop/src/main/sync/certificate-pinning.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import type tls from 'node:tls'
-import { DEFAULT_SYNC_CERT_HOSTNAME, PINNED_CERTIFICATE_HASHES_BY_HOST } from './certificate-pins'
+import { DEFAULT_SYNC_CERT_HOSTNAME, PINNED_CERTIFICATE_HASHES_BY_HOST } from '@memry/sync-client/certificate-pins'
 
 const mockApp = vi.hoisted(() => ({ isPackaged: false }))
