+++ packages/sync-client/src/cert-hash-cli.test.ts
