+++ packages/sync-client/src/cert-hash-cli.ts
