+++ packages/sync-client/src/certificate-pins.test.ts
